Ensure http_file load statement is at top of file.
We now take multiple passes to parse the original macro file,
first reading all load statements to see if any of them are the
`http_file` load statement we expect (these words are chosen
carefully—it will not notice a load statement other than the
one we expect even if it accomplishes the same goal). If none
of them were, then it will be added to the top of the file.

The managed block is considered separately in a second pass
through the file.

Closes #3.