Add option to ignore specific CVEs.
grype_test rules now accept an `ignore_cves` attribute
which represents a list of CVE strings, e.g.,
CVE-2021-33574, which should not be considered test
failures if they are discovered in the image under test.

Also, the documentation has intentionally NOT been updated
in this commit in order to verify the behavior of the CI
system, which should generate and commit updated documentation
if necessary.