Fix platform-specific issues for darwin and windows
On Windows, the grype configuration file is not correctly templated,
which causes the script generation to fail.

On all platforms other than linux-amd64, the Syft rule used the Grype
binary, which causes those builds to fail.

This patch resolves both of those issues to allow the rules to run on
all platforms.

Closes: #12